created home interface adapter and refactored HomeView to be consistent with other views
