+++ src/main/java/app/AppBuilder.java
@@ -4,12 +4,12 @@ import data_access.UserDataAccessObject;
 import entity.user.ClubUserFactory;
 import entity.user.StudentUserFactory;
 import interface_adapter.ViewManagerModel;
-import interface_adapter.data_access.HomeViewModel;
+import interface_adapter.home.HomeViewModel;
 import use_case.login.LoginInputBoundary;
 import use_case.login.LoginInteractor;
 import use_case.login.LoginOutputBoundary;
 import view.ClubSignupView;
-import view.HomeView.HomeView;
+import view.HomeView;
 import view.LoginView;
 import view.StudentSignupView;
 import view.ViewManager;
