--- src/main/java/interface_adapter/data_access/DataAccessPresenter.java
@@ -1,4 +1,0 @@
-package interface_adapter.data_access;
-
-public class DataAccessPresenter {
-}
