+++ src/main/java/interface_adapter/home/HomePresenter.java
@@ -1,0 +1,4 @@
+package interface_adapter.home;
+
+public class HomePresenter {
+}
