+++ src/main/java/interface_adapter/home/HomeState.java
@@ -1,0 +1,4 @@
+package interface_adapter.home;
+
+public class HomeState {
+}
