+++ src/main/java/interface_adapter/home/HomeViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.data_access;
+package interface_adapter.home;
 
 
 import interface_adapter.ViewModel;
